requirements.txt
# Core ML / NLP
sentence-transformers>=2.7.0
transformers>=4.40.0
torch>=2.2.0

# Vector stores
faiss-cpu>=1.8.0
chromadb>=0.5.0

# RAG framework
llama-index>=0.10.0
llama-index-llms-huggingface>=0.2.0
llama-index-embeddings-huggingface>=0.2.0

# Data handling
pandas>=2.2.0
numpy>=1.26.0
ijson>=3.2.0
datasets>=2.19.0

# Frontend
streamlit>=1.35.0

# Evaluation & visualization
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
huggingface_hub>=0.23.0
google-genai>=1.0.0

# Testing
pytest>=8.1.0

matplotlib
seaborn